requirements 2.txt
google-play-scraper==1.2.4
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
tqdm==4.66.1
pydantic==2.5.0
python-dotenv==1.0.0